requirements.txt
faiss_cpu
keras
numpy
scikit_learn
scipy
setuptools
tensorflow>=2.4.0
tensorflow_datasets
matplotlib
